+++ botx/clients/methods/base.py
@@ -1,19 +1,8 @@
+"""Definition for base class that is responsible for request to BotX API."""
+from __future__ import annotations
+
+import typing
 from abc import ABC, abstractmethod
-from typing import (
-    Any,
-    Awaitable,
-    Callable,
-    Dict,
-    Generic,
-    List,
-    Mapping,
-    NoReturn,
-    Optional,
-    Sequence,
-    Type,
-    TypeVar,
-    Union,
-)
 
 from httpx import URL, Response
 from pydantic import BaseConfig, BaseModel, Extra
@@ -23,32 +12,63 @@ from botx.clients.methods.request_wrapper import HTTPRequest, PrimitiveDataType
 from botx.models.enums import Statuses
 
 try:
-    from typing import Literal
+    from typing import Literal  # noqa: WPS433
 except ImportError:
-    from typing_extensions import Literal  # type: ignore
+    from typing_extensions import Literal  # noqa: WPS433, WPS440
 
 PRIMITIVES_FOR_QUERY = (str, int, float, bool, type(None))
 
-ResponseT = TypeVar("ResponseT")
-SyncErrorHandler = Callable[["BotXMethod", Response], NoReturn]
-AsyncErrorHandler = Callable[["BotXMethod", Response], Awaitable[NoReturn]]
-ErrorHandler = Union[SyncErrorHandler, AsyncErrorHandler]
-ErrorHandlersInMethod = Union[Sequence[ErrorHandler], ErrorHandler]
+ResponseT = typing.TypeVar("ResponseT")
+SyncErrorHandler = typing.Callable[["BotXMethod", Response], typing.NoReturn]
+AsyncErrorHandler = typing.Callable[
+    ["BotXMethod", Response], typing.Awaitable[typing.NoReturn],
+]
+ErrorHandler = typing.Union[SyncErrorHandler, AsyncErrorHandler]
+ErrorHandlersInMethod = typing.Union[typing.Sequence[ErrorHandler], ErrorHandler]
 
 
-class APIResponse(GenericModel, Generic[ResponseT]):
+def _convert_query_to_primitives(
+        query_params: typing.Mapping[str, typing.Any],
+) -> typing.Dict[str, PrimitiveDataType]:
+    converted_params = {}
+    for param_key, param_value in query_params.items():
+        if isinstance(param_value, PRIMITIVES_FOR_QUERY):
+            converted_params[param_key] = param_value
+        else:
+            converted_params[param_key] = str(param_value)
+
+    return converted_params
+
+
+class APIResponse(GenericModel, typing.Generic[ResponseT]):
+    """Model for successful response from BotX API."""
+
+    #: status of requested operation response.
     status: Literal[Statuses.ok] = Statuses.ok
+
+    #: generic response shape.
     result: ResponseT
 
 
-class APIErrorResponse(GenericModel, Generic[ResponseT]):
+class APIErrorResponse(GenericModel, typing.Generic[ResponseT]):
+    """Model for error response from BotX API."""
+
+    #: status of requested operation response.
     status: Literal[Statuses.error] = Statuses.error
+
+    #: reason why operation failed
     reason: str
-    errors: List[str]
+
+    #: errors from API.
+    errors: typing.List[str]
+
+    #: additional payload with more data about error.
     error_data: ResponseT
 
 
-class AbstractBotXMethod(ABC, Generic[ResponseT]):
+class AbstractBotXMethod(ABC, typing.Generic[ResponseT]):
+    """Abstract base class for BotX request."""
+
     @property
     @abstractmethod
     def __url__(self) -> str:
@@ -61,63 +81,74 @@ class AbstractBotXMethod(ABC, Generic[ResponseT]):
 
     @property
     @abstractmethod
-    def __returning__(self) -> Type[Any]:
+    def __returning__(self) -> typing.Type[typing.Any]:
         """Shape returned from method that can be parsed by pydantic."""
+
+    @property
+    def __errors_handlers__(self) -> typing.Dict[int, ErrorHandlersInMethod]:
+        """Special handlers for errors from BotX API by status code and handler."""
+        return {}
+
+    @property
+    def __result_extractor__(self) -> typing.Optional[
+        typing.Callable[[BotXMethod, typing.Any], ResponseT]
+    ]:
+        """Extractor for response shape from BotX API."""
+        return None
 
 
 CREDENTIALS_FIELDS = ("token", "host", "scheme")
 
 
 class BaseBotXMethod(AbstractBotXMethod[ResponseT], ABC):
-    host: str = ""
-    token: str = ""
-    scheme: str = "https"
+    """Base logic that is responsible for configuration and shortcuts for fields."""
 
-    def configure(self, *, host: str, token: str, scheme: str = "https") -> None:
-        self.token = token
-        self.host = host
-        self.scheme = scheme
+    #: host where request should be sent.
+    host: str = ""
+
+    #: token for request.
+    token: str = ""
+
+    #: HTTP scheme for request.
+    scheme: str = "https"
 
     @property
     def base_url(self) -> str:
+        """Base URL(scheme + hsot) for httpx clients."""
         return "{scheme}://{host}".format(scheme=self.scheme, host=self.host)
 
     @property
     def url(self) -> str:
+        """Full URL for request."""
         return str(URL(self.base_url).join(self.__url__))
 
     @property
     def http_method(self) -> str:
+        """HTTP method for request."""
         return self.__method__
 
     @property
-    def headers(self) -> Dict[str, str]:
+    def headers(self) -> typing.Dict[str, str]:
+        """Headers that should be used in request."""
         return {"Content-Type": "application/json"}
 
     @property
-    def params(self) -> Dict[str, PrimitiveDataType]:
+    def query_params(self) -> typing.Dict[str, PrimitiveDataType]:
+        """Query string query_params for request."""
         return {}
 
     @property
-    def __errors_handlers__(self) -> Dict[int, ErrorHandlersInMethod]:
-        return {}
-
-    @property
-    def error_handlers(self) -> Dict[int, ErrorHandlersInMethod]:
+    def error_handlers(self) -> typing.Dict[int, ErrorHandlersInMethod]:
         return self.__errors_handlers__
 
     @property
-    def __result_extractor__(
-        self,
-    ) -> Optional[Callable[["BotXMethod", Any], ResponseT]]:
-        return None
-
-    @property
-    def result_extractor(self) -> Optional[Callable[["BotXMethod", Any], ResponseT]]:
+    def result_extractor(
+            self,
+    ) -> typing.Optional[typing.Callable[[BotXMethod, typing.Any], ResponseT]]:
         return self.__result_extractor__
 
     @property
-    def returning(self) -> Type[Any]:
+    def returning(self) -> typing.Type[typing.Any]:
         return self.__returning__
 
 
@@ -128,11 +159,23 @@ class BotXMethod(BaseBotXMethod[ResponseT], BaseModel, ABC):
         arbitrary_types_allowed = True
         orm_mode = True
 
-    def encode(self) -> Optional[str]:
+    def configure(self, *, host: str, token: str, scheme: str = "https") -> None:
+        """Configure request with credentials and transport related stuff.
+
+        Arguments:
+            host: host where request should be sent.
+            token: token for request.
+            scheme: HTTP scheme for request.
+        """
+        self.token = token
+        self.host = host
+        self.scheme = scheme
+
+    def encode(self) -> typing.Optional[str]:
         return self.json(by_alias=True, exclude=set(CREDENTIALS_FIELDS))
 
     def build_http_request(self) -> HTTPRequest:
-        request_params = self.params
+        request_params = self.query_params
         request_data = self.encode()
 
         if self.__method__ == "GET":
@@ -156,20 +199,7 @@ class BotXMethod(BaseBotXMethod[ResponseT], BaseModel, ABC):
 
 class AuthorizedBotXMethod(BotXMethod[ResponseT], ABC):
     @property
-    def headers(self) -> Dict[str, str]:
+    def headers(self) -> typing.Dict[str, str]:
         headers = super().headers
         headers["Authorization"] = "Bearer {token}".format(token=self.token)
         return headers
-
-
-def _convert_query_to_primitives(
-    query_params: Mapping[str, Any],
-) -> Dict[str, PrimitiveDataType]:
-    converted_params = {}
-    for param_key, param_value in query_params.items():
-        if isinstance(param_value, PRIMITIVES_FOR_QUERY):
-            converted_params[param_key] = param_value
-        else:
-            converted_params[param_key] = str(param_value)
-
-    return converted_params