+++ botx/collecting/collectors/collector.py
@@ -57,7 +57,7 @@ class Collector(  # noqa: WPS215
         self._add_handlers(collector.handlers, dependencies)
 
     def command_for(self, *args: Any) -> str:
-        """Find handler and build a command string using passed body params.
+        """Find handler and build a command string using passed body query_params.
 
         Arguments:
             args: sequence of elements where first element should be name of handler.