+++ botx/middlewares/base.py
@@ -39,7 +39,7 @@ class BaseMiddleware:
     def __init__(
         self, executor: Executor, dispatch: Optional[MiddlewareDispatcher] = None,
     ) -> None:
-        """Init middleware with required params.
+        """Init middleware with required query_params.
 
         Arguments:
             executor: callable object that accept message and will be executed after