+++ botx/testing/botx_mock/asgi/routes/chats.py
@@ -12,8 +12,8 @@ from botx.clients.methods.v3.chats.info import Info
 from botx.clients.methods.v3.chats.remove_user import RemoveUser
 from botx.clients.methods.v3.chats.stealth_disable import StealthDisable
 from botx.clients.methods.v3.chats.stealth_set import StealthSet
-from botx.models.chats import ChatFromSearch
 from botx.clients.types.response_results import ChatCreatedResult
+from botx.models.chats import ChatFromSearch
 from botx.models.events import UserInChatCreated
 from botx.testing.botx_mock.asgi.binders import bind_implementation_to_method
 from botx.testing.botx_mock.asgi.messages import add_request_to_collection